+++ src/mf_utilities.h
@@ -4,6 +4,6 @@
 #include "Openflow/types.h"
 
 uint32_t generate_random();
-void inverse_memcpy(void* dst, void* src, uint8_t len);
+void inverse_memcpy(void* dst, void* src, uint16_t len);
 
 #endif
